+++ KrieptoBot.Application/Settings/TradingSettings.cs
@@ -10,8 +10,8 @@ namespace KrieptoBot.Application.Settings
         public string Interval { get; set; }
         public IEnumerable<string> MarketsToWatch { get; set;}
         public DateTime CurrentTime { get; set; }
-        public int BuyMargin { get; set;}
-        public int SellMargin { get;  set;}
+        public decimal BuyMargin { get; set;}
+        public decimal SellMargin { get;  set;}
         public bool IsSimulation { get;  set;}
     }
 }
